+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 All notable changes to the "mjml" extension will be documented in this file.
 
-### [0.0.1]
-This is the initial release of this extension.
+### [0.0.4] (2017-06-21)
+* [#1](https://github.com/attilabuti/vscode-mjml/pull/1): fixed image path issue.
+* [#2](https://github.com/attilabuti/vscode-mjml/issues/2): fixed mj-include issue.
+* [new] Configuration property `mjml.preserveFocus`: preserve focus of Text Editor after preview open.
+* [new] Configuration property `mjml.updateWhenTyping`: update preview when typing.
+* `MJML: Open Preview` was renamed to `MJML: Open Preview to the Side`
+* `MJML: Generate HTML` was renamed to `MJML: Export HTML`
+
+### [0.0.2] (2017-05-08)
+* Some fixes.
+
+### [0.0.1] (2017-05-07)
+* This is the initial release of this extension.